+++ db/access.php
@@ -26,8 +26,13 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->version   = 2017110802;        // The current plugin version (Date: YYYYMMDDXX).
-$plugin->requires = 2017051500.00; // This is Moodle 3.3 (Build: 20170515).
-$plugin->maturity = MATURITY_BETA;
-$plugin->component = 'format_ned';
-$plugin->release = '3.3.1.3';
+$capabilities = array(
+    'format/ned:formatupdate' => array(
+        'captype' => 'write',
+        'contextlevel' => CONTEXT_COURSE,
+        'archetypes' => array(
+            'editingteacher' => CAP_ALLOW,
+            'manager' => CAP_ALLOW
+        )
+    )
+);